Implement setPath path following.
I have modified the path follower to take a single SetPath command, rather than having to set a line and setting a curve each time. The array of points passed in to the command will determine which gets called. The first item in the array is always the x,y of the destination. If it is a straight line, the second (and last) item should be the starting point (though in reality the starting point is ignored by the code). If it is to be a curve, the first point is the destination. The second is the closest to the end control point. the third is the closest to the beginning control point, and the fourth is the beginning point. Again, the starting point is ignored by the code - the path follower always starts with where the robot is now. The second parameter is the angle you want the robot facing at destination.

I have also made an overload method. If you add a third paramater MoveType, you can add RAM, LOOSE, PRECISE to speed up movements, make them more precise at the target, etc. These are carryovers of prior code - I don't see us needing them this year (maybe precise if we have shooting trouble);

Finally, we will retain a separate "turnOnly" method to quickly turn the robot. You can call it directly, or call setPath(targetAngle) without any other parameter. There is also a MoveType LOOSE overload for a fast turn.